Support write-only meta-properties
According to the discussion in #3: "Since Java Beans v1.0 allows write-only properties, I suspect we have to allow for them, but not beyond the basic possibility of an exception. Adding an isWriteOnly() method seems like overkill." (jodastephen) 'AbstractMetaProperty' now supports write-only properties according to 'MetaProperty.get(Object)'.

Additionally the order of the Boolean constructor arguments was changed to reflect the order of the methods in 'MetaProperty'.