Move reusable parts of 'FieldMetaPropertyTest' into own class
Most of the code from the test class is chaining factory methods and only one method actually creates instances. Only the last one is specific to this test; the chaining methods can be reused for every test class which can create a 'MetaProperty' based on the name of a field. Hence an additional abstract class was introduced which only leaves the method which creates an instance for a field name abstract.