Reduce strictness of annotations test in 'AbstractMetaPropertyTest'.
It became obvious that the general test for the implementation of 'MetaProperty.annotations()' was still too strict as it required exactly one annotation of the correct type. But the straight forward implementation of 'MethodMetaProperty.annotations()' would return two instances because it queried the get and set methods for annotations. Only having one annotation does not look very promising because there seems to be no way to place it such that all imaginable implementations would return it.

Instead the test was softened to only expect at least one annotation (and only ones of the correct type).